Added User Profile and fixed authentication
Added user profile, now you can click the user icon on the navbar to go
to user profile. [WIP]
Fixed Issue with authentication where sign-in should be a query not a
mutation. Made it so that login and signup only return JWT-Token. No
user info. User info is only gotten once the token is decoded.